Refactor: Update NFCTag model and related code
